Deployed to heroku and added link to README
Added basic streamlit app pages so I could perform
an early deployment.

Had to change the python stack in heroku from
heroku-22 to heroku-20 using the following command:
heroku stack:set heroku-20 -a pp5-house-price-predictor

Initial warning about slug size but app is running
as expect and can be accessed.

Another warning regarding soon to be deprecated
heroku-20 stack but this can be addressed later as
this was only to test deployment.

End of life for heroku-20 is 30th April 2025.
Heroku provided the following links for more information:
https://devcenter.heroku.com/articles/upgrading-to-the-latest-stack
https://help.heroku.com/NPN275RK/heroku-20-end-of-life-faq

App is running at:
https://pp5-house-price-predictor-3b886bb0d166.herokuapp.com/